+++ protected/humhub/modules/admin/notifications/IncludeGroupNotification.php
@@ -29,7 +29,7 @@ class IncludeGroupNotification extends BaseNotification
      */
     public function getUrl()
     {
-        return Url::to(['/directory/directory/groups']);
+        return Url::to(['/user/people']);
     }
 
     /**